extract example transition
